+++ test/env.js
@@ -12,6 +12,11 @@ const config = {
 };
 
 describe('Env layer', () => {
+  after(() => {
+    delete process.env.some_data_here;
+    delete process.env.my_prefix_here;
+  });
+
   it('should parse environment and use path', async () => {
     process.env.some_data_here = JSON.stringify(config);
     const onion = new Onion();
@@ -37,6 +42,22 @@ describe('Env layer', () => {
     });
   });
 
+  it('should normalize prefix, if ends with parsingSeparator', async () => {
+    const config = {
+      aNumber: 1,
+      aString: 'hello',
+      someObject: {
+        dozed: 12,
+        nothingHere: false,
+      }
+    };
+    process.env.my_prefix_here = JSON.stringify(config);
+    const onion = new Onion({});
+    const layer = new Onion.LAYERS.Env({ prefix: 'my_prefix_' });
+    await onion.addLayer(layer);
+    layer.options.prefix.should.be.eql('my_prefix');
+  });
+
   it('should use prefix', async () => {
     const config = {
       aNumber: 1,
@@ -46,21 +67,13 @@ describe('Env layer', () => {
         nothingHere: false,
       }
     };
-    process.env.some_data_here = JSON.stringify(config);
+    process.env.my_prefix_here = JSON.stringify(config);
     const onion = new Onion({});
-    await onion.addLayer(new Onion.LAYERS.Env({ prefix: 'some_data_' }));
+    await onion.addLayer(new Onion.LAYERS.Env({ prefix: 'my_prefix' }));
     onion.get().should.be.eql({
-      some: {
-        data: {
-          here: JSON.stringify(config),
-        },
-      },
+      here: JSON.stringify(config),
     });
-    onion.get('some').should.be.eql({
-      data: {
-        here: JSON.stringify(config),
-      },
-    });
+    onion.get('here').should.be.eql(JSON.stringify(config));
   });
 
   it('should parse JSON in environment', async () => {
@@ -76,11 +89,7 @@ describe('Env layer', () => {
     const onion = new Onion({});
     await onion.addLayer(new Onion.LAYERS.Env({ prefix: 'some_data_', json: true }));
     onion.get().should.be.eql({
-      some: {
-        data: {
-          here: config,
-        },
-      },
+      here: config,
     });
   });
 });